task1/requirements.txt
nats-py==2.8.0
python-dotenv==1.0.0
mysql-connector-python==8.4.0